requirements.txt
python-telegram-bot==13.15
APScheduler==3.6.3
tornado==6.1
pytz==2025.2
tzlocal==5.3.1
certifi==2025.4.26
cachetools==4.2.2
six==1.17.0
